requirements.txt
faster-whisper
pyannote.audio
torch
torchaudio
sentence-transformers
numpy
openai
python-dotenv
requests
fastapi
uvicorn
python-multipart
tqdm
feedparser
openai
librosa
soundfile
